Pass parameters directly to Injector::instantiate() and invoke()
There is no reason to use ControllerValueResolver to handle these named parameters, and this had the side effect to have the named parameters used to resolve injected properties and method parameters as well (in addition to the constructor), which was confusing at best.

The parameters coming from route match (and plugins) should only be used to resolve class constructors and controller methods, not injected methods & properties.